spring security / implementation variable d'environnement

+++ src/main/java/edu/fprs/jfb_2025_demo/controller/ProduitController.java
@@ -3,12 +3,15 @@ package edu.fprs.jfb_2025_demo.controller;
 import edu.fprs.jfb_2025_demo.dao.ProduitDao;
 import edu.fprs.jfb_2025_demo.model.Etat;
 import edu.fprs.jfb_2025_demo.model.Produit;
+import edu.fprs.jfb_2025_demo.security.AppUserDetails;
 import edu.fprs.jfb_2025_demo.security.IsRedacteur;
 import edu.fprs.jfb_2025_demo.security.IsUtilisateur;
+import edu.fprs.jfb_2025_demo.security.SecuriteUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,16 +22,14 @@ import java.util.Optional;
 @RestController
 public class ProduitController {
 
-//    @Autowired
-//    protected ProduitDao produitDao;
-
     protected ProduitDao produitDao;
+    protected SecuriteUtils securiteUtils;
 
     @Autowired
-    public ProduitController(ProduitDao produitDao) {
+    public ProduitController(ProduitDao produitDao, SecuriteUtils securiteUtils) {
         this.produitDao = produitDao;
+        this.securiteUtils = securiteUtils;
     }
-
 
     @GetMapping("/produit/{id}")
     public ResponseEntity<Produit> get(@PathVariable int id) {
@@ -51,7 +52,11 @@ public class ProduitController {
     }
 
     @PostMapping("/produit")
-    public ResponseEntity<Produit> save(@RequestBody Produit produit) {
+    @IsUtilisateur
+    public ResponseEntity<Produit> save(@RequestBody @Valid Produit produit,
+                                        @AuthenticationPrincipal AppUserDetails userDetails) {
+
+        produit.setCreateur(userDetails.getUtilisateur());
 
         //si le produit recu n'a pas d'etat alors on indiquide qu'il neuf par defaut
         if (produit.getEtat() == null) {
@@ -68,12 +73,24 @@ public class ProduitController {
 
     @DeleteMapping("/produit/{id}")
     @IsRedacteur
-    public ResponseEntity<Produit> delete(@PathVariable int id) {
+    @IsUtilisateur
+    public ResponseEntity<Produit> delete(
+            @PathVariable int id,
+            @AuthenticationPrincipal AppUserDetails userDetails) {
 
         Optional<Produit> optionalProduit = produitDao.findById(id);
 
         if (optionalProduit.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        //récupérer le role de la personne
+        String role = securiteUtils.getRole(userDetails);
+
+        //si l'id createur du produit est different de l'id de la personnne connectee
+        if (!role.equals("ROLE_ADMINISTRATEUR") ||
+                optionalProduit.get().getCreateur().getId() != userDetails.getUtilisateur().getId()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
 
         produitDao.deleteById(id);
@@ -83,6 +100,7 @@ public class ProduitController {
     }
 
     @PutMapping("/produit/{id}")
+    @IsUtilisateur
     public ResponseEntity<Produit> update(
             @PathVariable int id,
             @RequestBody @Valid Produit produit) {
